Home Navigation - Chat UI
- redesigned the home page to match the desired
    layout
- added a new chat page for user interactions
